+++ projectroles/tests/test_permissions_ajax.py
@@ -213,6 +213,22 @@ class TestProjectViews(TestProjectPermissionBase):
         self.project.set_public()
         self.assert_response(url, self.anonymous, 401, method='POST')
 
+    def test_current_user(self):
+        """Test CurrentUserRetrieveAjaxView access"""
+        url = reverse('projectroles:ajax_user_current')
+        good_users = [
+            self.superuser,
+            self.owner_as.user,
+            self.delegate_as.user,
+            self.contributor_as.user,
+            self.guest_as.user,
+            self.user_no_roles,
+        ]
+        bad_users = [self.anonymous]
+        self.assert_response(url, good_users, 200)
+        self.assert_response(url, bad_users, 403)
+
+
     @override_settings(PROJECTROLES_ALLOW_LOCAL_USERS=True)
     def test_user_autocomplete_ajax(self):
         """Test UserAutocompleteAjaxView access"""
@@ -226,6 +242,5 @@ class TestProjectViews(TestProjectPermissionBase):
             self.user_no_roles,
         ]
         bad_users = [self.anonymous]
-
         self.assert_response(url, good_users, 200)
         self.assert_response(url, bad_users, 403)